requirements.txt
# Core Dependencies
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.17.0
python-dotenv>=1.0.0
requests>=2.31.0

# AI/LLM
langchain>=0.1.0
langchain-google-genai>=1.0.0
google-generativeai>=0.3.0

# Testing
pytest>=7.4.0
pytest-mock>=3.12.0

# Development
black>=23.0.0
flake8>=6.1.0
